Template: Split Header into more components
https://phabricator.endlessm.com/T31510
