+++ template-ui/src/components/HeaderNavBarBrand.vue
@@ -1,0 +1,21 @@
+<template>
+  <b-navbar-brand to="/">
+    <img
+      v-if="channel.thumbnail"
+      :src="channel.thumbnail"
+      :alt="channel.title"
+    />
+    <span v-else>{{ channel.title }}</span>
+  </b-navbar-brand>
+</template>
+
+<script>
+import { mapState } from 'vuex';
+
+export default {
+  name: 'HeaderNavBarBrand',
+  computed: {
+    ...mapState(['channel']),
+  },
+};
+</script>
